+++ ARatmospy/psds/sharcstemppsds.py
@@ -34,7 +34,7 @@ d = pscale * m
 per_len = perlen
 f = np.arange(per_len)
 
-hz = np.roll(f - per_len / 2, np.int(per_len / 2)) / per_len * rate
+hz = np.roll(f - per_len / 2, np.int64(per_len / 2)) / per_len * rate
 shz = np.sort(hz)
 omega = 2 * np.pi * shz / rate
 ahz = np.argsort(hz)
@@ -77,7 +77,7 @@ while fmode is not "q":
         continue
 
     print("Analyzing ", fname)
-    psd = np.memmap(fname, dtype="float64", mode="r", shape=(perlen, bign, bign))
+    psd = np.memmap(fname, dtype=np.float64, mode="r", shape=(perlen, bign, bign))
     # fnum = fname.split('_')[1].split('-')[0]
 
     mode = "2,1"
